Beginning to add d14rs
I'm really not liking the "rails"-style directory structure.

Now that I'm beginning to have d14rs I'm beginning to need some of the
business logic/game-ified stuff and I'm having trouble finding the right
place to plug in the impure parts, that will keep the task of ranking
from feeling like a chore.

I've read a lot of posts about how to best go about writing a redux app
that's slightly more complicated than can be delivered with the example
docs. As part of that I've learned that the shape of the state tree
should be pretty well abstracted from the rest of the application. Since
you use reducers to turn actions into state mutations, you should also
use selectors so no individual part of the UI (or you coding it) has to
care much about where your state is stored, so long as it is, and you
can deterministically get it back.

The docs suggested method of keeping your "containers" and your
"components" separate makes sense from a code standard. It's useful to
not tightly couple how you get your data to how you render the data in a
UI, so you can write tests around your storage of the data, and write
tests around rendering your UI. But, whether a node knows how to map
state, or dumbly works on props, is an implementation for THAT
component. No one up the tree, or down the tree, should have to care.
But in working on this commit, I've been bitten every time I changed a
component to be wrapped in a container, because watch updates, and I
refresh in the browser and nothing changes in the render. Going and
changing the include path for anywhere that renders a component will end
up more painful than the reminder to separate concerns is worth.

So, I think I'm going to migrate to module/domain/feature style
organization. But each component will use default export to declare
itself, and inside the file will determine whether that export is a
regular react component, or a connected one. I'm going to try to export
a destructurable { Component } so that I can write UI tests on just the
component so I don't have to find some clever way to intercept the state
mapping step, and just pass props in instead.

From reading the conversation in this pull request:
https://github.com/reactjs/redux/issues/1171
and this blog post:
http://randycoulman.com/blog/2016/09/27/modular-reducers-and-selectors/
I'm frustrated that somehow module organization means that you disregard
the recommendation to not pair your reducers and action creators. Since
the purpose of using reducers and selectors is to abstract away the
shape of the state tree as an implementation detail, it sounds to me
like it's appropriate to have a "state"/"reducer" module that handles
building and accessing your state. This may just be hubris, and I may
come back admitting my faults and wrongs.