labeling connected subgraphs to make everything easier
With labeled subgraphs I can show the ranking of connected subgraphs
and make better decisions when picking cards out of the deck for
ranking.

I fp-ed hard tonight. Not entirely happy with how branch-y it is right
now. But, I don't expect this solution (graph labeling) to live long. I'm just proud of
it right now.

Rankerrater is going to need to be a lot more cognizant of making chains
rather than graphs, so labeling will probably not be a helpful solution
as Rankerrater gets better at building rankings with fewer attempts at rank
description.